+++ lib/wasmify/rails/builder.rb
@@ -9,6 +9,8 @@ module Wasmify
   module Rails
     # A wrapper for rbwasm build command
     class Builder
+      ORIGINAL_EXCLUDED_GEMS = RubyWasm::Packager::EXCLUDED_GEMS.dup.freeze
+
       attr_reader :output_dir, :target
 
       def initialize(output_dir: Wasmify::Rails.config.tmp_dir, target: Wasmify::Rails.config.wasm_target)
@@ -16,16 +18,28 @@ module Wasmify
         @target = target
       end
 
-      def run(name:, exclude_gems: [], opts: "")
+      def run(name:, exclude_gems: [], ignore_extensions: [], opts: "")
+        # Reset excluded gems
+        RubyWasm::Packager::EXCLUDED_GEMS.replace(ORIGINAL_EXCLUDED_GEMS)
+        $exclude_gems = []
         $exclude_exts = []
 
         # Add configured excluded gems
         Wasmify::Rails.config.exclude_gems.each do |gem_name|
-          $exclude_exts << gem_name
+          RubyWasm::Packager::EXCLUDED_GEMS << gem_name
+        end
+
+        # Add configured excluded extensions
+        Wasmify::Rails.config.ignore_gem_extensions.each do |ext_name|
+          $exclude_exts << ext_name
         end
 
         # Add additional excluded gems
         exclude_gems.each do |gem_name|
+          RubyWasm::Packager::EXCLUDED_GEMS << gem_name
+        end
+
+        ignore_extensions.each do |gem_name|
           $exclude_exts << gem_name
         end
 
@@ -46,7 +60,7 @@ module Wasmify
           -o #{File.join(output_dir, name)}
         ) + opts
 
-        patches_dir = ::Rails.root.join("ruby_wasm_patches").to_s
+        patches_dir = Wasmify::Rails.config.root_dir.join("ruby_wasm_patches").to_s
 
         if File.directory?(patches_dir)
           Dir.children(patches_dir).each do |patch|