+++ lib/wasmify/rails/packer.rb
@@ -12,7 +12,24 @@ module Wasmify
 
       attr_reader :output_dir
 
-      include ActionView::Helpers::NumberHelper
+      if defined?(ActionView::Helpers::NumberHelper)
+        include ActionView::Helpers::NumberHelper
+      else
+        include(Module.new do
+          def number_to_human_size(num)
+            num = num.to_i
+            human_size = case
+            when num < 1024
+              "#{num} bytes"
+            when num < 1024*1024
+              "#{(num.to_f/1024).round(1)} KB"
+            else
+              "#{(num.to_f/1024/1024).round(1)} MB"
+            end
+            human_size
+          end
+        end)
+      end
 
       def initialize(output_dir: Wasmify::Rails.config.output_dir)
         @output_dir = output_dir
@@ -24,12 +41,14 @@ module Wasmify
                "Please see installations instructions at: https://github.com/kateinoigakukun/wasi-vfs"
         end
 
+        pack_root = Wasmify::Rails.config.pack_root
+
         args = %W[
           pack #{ruby_wasm_path}
         ]
 
         directories.each do |dir|
-          args << "--dir #{::Rails.root.join(dir)}::/rails/#{dir}"
+          args << "--dir #{Wasmify::Rails.config.root_dir.join(dir)}::#{pack_root}/#{dir}"
         end
 
         output_path = File.join(output_dir, name)
@@ -38,14 +57,14 @@ module Wasmify
 
         # Generate a temporary directory with the require root files
         Dir.mktmpdir do |tdir|
-          ROOT_FILES.each do |file|
-            FileUtils.cp(::Rails.root.join(file), tdir) if File.exist?(::Rails.root.join(file))
+          (ROOT_FILES + (Wasmify::Rails.config.additional_root_files || [])).each do |file|
+            FileUtils.cp(Wasmify::Rails.config.root_dir.join(file), tdir) if File.exist?(Wasmify::Rails.config.root_dir.join(file))
           end
 
           # Create a storage/ directory for Active Storage attachments
           FileUtils.mkdir_p(File.join(tdir, storage_dir)) if storage_dir
 
-          args << "--dir #{tdir}::/rails"
+          args << "--dir #{tdir}::#{pack_root}"
 
           args << "-o #{output_path}"