+++ lib/wasmify/rails/railtie.rb
@@ -15,3 +15,15 @@ module Wasmify
     end
   end
 end
+
+
+require "action_mailer/null_delivery"
+
+# NullDB for Active Record
+ActiveRecord::ConnectionAdapters.register("nulldb", "ActiveRecord::ConnectionAdapters::NullDBAdapter", "active_record/connection_adapters/nulldb_adapter")
+
+# SQLite3 Wasm adapter
+ActiveRecord::ConnectionAdapters.register("sqlite3_wasm", "ActiveRecord::ConnectionAdapters::SQLite3WasmAdapter", "active_record/connection_adapters/sqlite3_wasm_adapter")
+
+# PGlite adapter
+ActiveRecord::ConnectionAdapters.register("pglite", "ActiveRecord::ConnectionAdapters::PGliteAdapter", "active_record/connection_adapters/pglite_adapter")